# openapi.yaml
openapi: 3.0.3
info:
  title: Polo Sanca Refrigeration Monitoring API
  description: |
    API for monitoring and managing refrigeration equipment across multiple companies and branches.

    ## Authentication
    - **User Authentication**: JWT tokens via `Authorization: Bearer <token>` header
    - **Equipment Authentication**: API keys via `X-API-Key: <key>` header

    ## Versioning
    - Current version: v1
    - Version included in URL path
    - Previous versions supported for 12 months after new release

  version: 1.0.0
  contact:
    name: Polo Sanca API Support
    email: api@polosanca.com
  license:
    name: Proprietary
servers:
  - url: https://api.polosanca.com/v1
    description: Production server
  - url: https://api-staging.polosanca.com/v1
    description: Staging server
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: List users
      description: Get list of users filtered by permissions (Company Admin or Global Admin)
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: company_id
          in: query
          description: Filter by company (Global Admin only)
          schema:
            type: string
        - name: role
          in: query
          description: Filter by role
          schema:
            type: string
            enum: [global_admin, company_admin, company_viewer]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
  /users/{user_id}:
    patch:
      tags:
        - Users
      summary: Update user
      description: Update user details or permissions
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [company_admin, company_viewer]
                branch_access:
                  $ref: '#/components/schemas/BranchAccess'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

    delete:
      tags:
        - Users
      summary: Deactivate user
      description: Deactivate or remove user
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deactivated successfully
...
  /equipments/telemetry:
    post:
      tags:
        - Telemetry
      summary: Submit equipment telemetry data
      description: Equipment posts monitoring data via this endpoint
      operationId: submitTelemetry
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryData'
      responses:
        '200':
          description: Data received successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  record_id:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /equipments/{equipment_id}/telemetry:
    get:
      tags:
        - Telemetry
      summary: Get historical telemetry data
      description: Retrieve historical monitoring data for equipment
      operationId: getEquipmentTelemetry
      security:
        - bearerAuth: []
      parameters:
        - name: equipment_id
          in: path
          required: true
          schema:
            type: string
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: interval
          in: query
          description: Data aggregation interval
          schema:
            type: string
            enum: [raw, 5min, 1hour, 1day]
            default: raw
        - name: fields
          in: query
          description: Comma-separated list of fields to include
          schema:
            type: string
            example: temperature,pressure,door
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 1000
          description: Higher limit for telemetry data
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  equipment_id:
                    type: string
                  serial:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TelemetryDataPoint'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Equipment API key for telemetry submission
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [global_admin, company_admin, company_viewer]
        company_id:
          type: string
        branch_access:
          $ref: '#/components/schemas/BranchAccess'
        status:
          type: string
          enum: [active, inactive, pending]
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
    BranchAccess:
      type: object
      properties:
        type:
          type: string
          enum: [full, restricted]
        branches:
          type: array
          items:
            type: string
          description: List of branch IDs (only for restricted type)
...
    TelemetryData:
      type: object
      required:
        - serial
        - temperature
        - pressure
        - door
        - heater
        - compressor
        - fan
      properties:
        serial:
          type: string
          example: EQ-12345
        temperature:
          type: number
          format: float
          example: 4.5
        pressure:
          type: number
          format: float
          example: 120.3
        door:
          type: integer
          enum: [0, 1]
          description: 0=closed, 1=open
        heater:
          type: integer
          enum: [0, 1]
          description: 0=off, 1=on
        compressor:
          type: integer
          enum: [0, 1]
          description: 0=off, 1=on
        fan:
          type: integer
          enum: [0, 1]
          description: 0=off, 1=on
    TelemetryDataPoint:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        temperature:
          type: number
          format: float
        pressure:
          type: number
          format: float
        door:
          type: integer
        heater:
          type: integer
        compressor:
          type: integer
        fan:
          type: integer
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
        total:
          type: integer
          description: Total number of items
        total_pages:
          type: integer
          description: Total number of pages
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad Request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - Authentication required or failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Not Found - Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...